HHH-1724 updated test which expected an Integer instead of Long as row count
git-svn-id: https://svn.jboss.org/repos/hibernate/core/trunk@17877 1b8cb986-b30d-0410-93ca-fae66ebed9b2

+++ testsuite/src/test/java/org/hibernate/test/idprops/IdentifierPropertyReferencesTest.java
@@ -182,6 +182,6 @@ public class IdentifierPropertyReferencesTest extends FunctionalTestCase {
 	}
 
 	private long extractCount(Criteria crit) {
-		return ( ( Integer ) crit.list().get( 0 ) ).intValue();
+		return ( ( Long ) crit.list().get( 0 ) ).longValue();
 	}
 }
